Move EnqueueMessages for BRS-021/023
