+++ source/B2BApi/Functions/EnqueueMessages/BRS_021_023/Activities/EnqueueEnergyResultsForBalanceResponsiblesActivity.cs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Model;
-using Energinet.DataHub.EDI.MasterData.Interfaces;
+using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
 using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Databricks.EnergyResults.Queries;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces.Models.EnergyResultMessages;
@@ -22,28 +21,27 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using EventId = Energinet.DataHub.EDI.BuildingBlocks.Domain.Models.EventId;
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Activities;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Activities;
 
 /// <summary>
-/// Enqueue energy results for Grid Area Owners as outgoing messages for the given calculation id.
+/// Enqueue energy results for Balance Responsibles as outgoing messages for the given calculation id.
 /// </summary>
-public class EnqueueEnergyResultsForGridAreaOwnersActivity(
-    ILogger<EnqueueEnergyResultsForGridAreaOwnersActivity> logger,
+public class EnqueueEnergyResultsForBalanceResponsiblesActivity(
+    ILogger<EnqueueEnergyResultsForBalanceResponsiblesActivity> logger,
     IServiceScopeFactory serviceScopeFactory,
     EnergyResultEnumerator energyResultEnumerator)
     : EnqueueEnergyResultsBaseActivity(logger, serviceScopeFactory, energyResultEnumerator)
 {
-    private readonly ILogger<EnqueueEnergyResultsForGridAreaOwnersActivity> _logger = logger;
+    private readonly ILogger<EnqueueEnergyResultsForBalanceResponsiblesActivity> _logger = logger;
     private readonly EnergyResultEnumerator _energyResultEnumerator = energyResultEnumerator;
 
-    [Function(nameof(EnqueueEnergyResultsForGridAreaOwnersActivity))]
+    [Function(nameof(EnqueueEnergyResultsForBalanceResponsiblesActivity))]
     public Task<int> Run(
         [ActivityTrigger] EnqueueMessagesInput input)
     {
-        var query = new EnergyResultPerGridAreaQuery(
+        var query = new EnergyResultPerBalanceResponsiblePerGridAreaQuery(
             _logger,
             _energyResultEnumerator.EdiDatabricksOptions,
-            input.GridAreaOwners,
             EventId.From(input.EventId),
             input.CalculationId);
 
@@ -52,9 +50,10 @@ public class EnqueueEnergyResultsForGridAreaOwnersActivity(
 
     protected override Task EnqueueAndCommitEnergyResult<TOutgoingMessage>(IOutgoingMessagesClient outgoingMessagesClient, TOutgoingMessage outgoingMessageDto)
     {
-        if (outgoingMessageDto is not EnergyResultPerGridAreaMessageDto perGridAreaMessageDto)
-            throw new ArgumentException($"The outgoing message dto is not of the expected type {typeof(EnergyResultPerGridAreaMessageDto).FullName}", nameof(outgoingMessageDto));
-
-        return outgoingMessagesClient.EnqueueAndCommitAsync(perGridAreaMessageDto, CancellationToken.None);
+        return outgoingMessageDto is EnergyResultPerBalanceResponsibleMessageDto perBalanceResponsibleMessageDto
+            ? outgoingMessagesClient.EnqueueAndCommitAsync(perBalanceResponsibleMessageDto, CancellationToken.None)
+            : throw new ArgumentException(
+                $"The outgoing message dto is not of the expected type {typeof(EnergyResultPerBalanceResponsibleMessageDto).FullName}",
+                nameof(outgoingMessageDto));
     }
 }
