+++ source/B2BApi/Functions/EnqueueMessages/BRS_021_023/Activities/EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity.cs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Model;
-using Energinet.DataHub.EDI.MasterData.Interfaces;
+using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
 using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Databricks.WholesaleResults.Queries;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces.Models.WholesaleResultMessages;
@@ -23,37 +22,34 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using EventId = Energinet.DataHub.EDI.BuildingBlocks.Domain.Models.EventId;
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Activities;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Activities;
 
-/// <summary>
-/// Enqueue wholesale results for Amount Per Charge to Energy Supplier and ChargeOwner as outgoing messages for the given calculation id.
-/// </summary>
-public class EnqueueWholesaleResultsForAmountPerChargesActivity(
-    ILogger<EnqueueWholesaleResultsForAmountPerChargesActivity> logger,
+public class EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity(
+    ILogger<EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity> logger,
     IServiceScopeFactory serviceScopeFactory,
     WholesaleResultEnumerator wholesaleResultEnumerator)
     : EnqueueWholesaleResultsBaseActivity(logger, serviceScopeFactory, wholesaleResultEnumerator)
 {
-    private readonly ILogger<EnqueueWholesaleResultsForAmountPerChargesActivity> _logger = logger;
+    private readonly ILogger<EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity> _logger = logger;
     private readonly WholesaleResultEnumerator _wholesaleResultEnumerator = wholesaleResultEnumerator;
 
     /// <summary>
-    /// Start an EnqueueWholesaleResultsForAmountPerCharges activity.
+    /// Start an EnqueueWholesaleResultsForMonthlyAmountPerCharges activity.
     /// <remarks>The <paramref name="input"/> type and return type must be that same as the <see cref="Run"/> method</remarks>
     /// <remarks>Changing the <paramref name="input"/> or return type might break the Durable Function's deserialization</remarks>
     /// </summary>
     public static Task<int> StartActivityAsync(EnqueueMessagesForActorInput input, TaskOrchestrationContext context, TaskOptions? options)
     {
         return context.CallActivityAsync<int>(
-            nameof(EnqueueWholesaleResultsForAmountPerChargesActivity),
+            nameof(EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity),
             input,
             options: options);
     }
 
-    [Function(nameof(EnqueueWholesaleResultsForAmountPerChargesActivity))]
+    [Function(nameof(EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity))]
     public Task<int> Run([ActivityTrigger] EnqueueMessagesForActorInput input)
     {
-        var query = new WholesaleAmountPerChargeQuery(
+        var query = new WholesaleMonthlyAmountPerChargeQuery(
             _logger,
             _wholesaleResultEnumerator.EdiDatabricksOptions,
             input.GridAreaOwners,
@@ -66,10 +62,10 @@ public class EnqueueWholesaleResultsForAmountPerChargesActivity(
 
     protected override Task EnqueueAndCommitWholesaleResult<TOutgoingMessage>(IOutgoingMessagesClient outgoingMessagesClient, TOutgoingMessage outgoingMessageDto)
     {
-        return outgoingMessageDto is WholesaleAmountPerChargeMessageDto amountPerChargeMessageDto
-            ? outgoingMessagesClient.EnqueueAndCommitAsync(amountPerChargeMessageDto, CancellationToken.None)
+        return outgoingMessageDto is WholesaleMonthlyAmountPerChargeMessageDto monthlyAmountPerChargeMessageDto
+            ? outgoingMessagesClient.EnqueueAndCommitAsync(monthlyAmountPerChargeMessageDto, CancellationToken.None)
             : throw new ArgumentException(
-                $"The outgoing message dto is not of the expected type {typeof(WholesaleAmountPerChargeMessageDto).FullName}",
+                $"The outgoing message dto is not of the expected type {typeof(WholesaleMonthlyAmountPerChargeMessageDto).FullName}",
                 nameof(outgoingMessageDto));
     }
 }
