+++ source/B2BApi/Functions/EnqueueMessages/BRS_021_023/Activities/EnqueueWholesaleResultsForTotalAmountsActivity.cs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Model;
-using Energinet.DataHub.EDI.MasterData.Interfaces;
+using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
 using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Databricks.WholesaleResults.Queries;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces.Models.WholesaleResultMessages;
@@ -23,34 +22,34 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using EventId = Energinet.DataHub.EDI.BuildingBlocks.Domain.Models.EventId;
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Activities;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Activities;
 
-public class EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity(
-    ILogger<EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity> logger,
+public class EnqueueWholesaleResultsForTotalAmountsActivity(
+    ILogger<EnqueueWholesaleResultsForTotalAmountsActivity> logger,
     IServiceScopeFactory serviceScopeFactory,
     WholesaleResultEnumerator wholesaleResultEnumerator)
     : EnqueueWholesaleResultsBaseActivity(logger, serviceScopeFactory, wholesaleResultEnumerator)
 {
-    private readonly ILogger<EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity> _logger = logger;
+    private readonly ILogger<EnqueueWholesaleResultsForTotalAmountsActivity> _logger = logger;
     private readonly WholesaleResultEnumerator _wholesaleResultEnumerator = wholesaleResultEnumerator;
 
     /// <summary>
-    /// Start an EnqueueWholesaleResultsForMonthlyAmountPerCharges activity.
+    /// Start an EnqueueWholesaleResultsForTotalAmountPerCharges activity.
     /// <remarks>The <paramref name="input"/> type and return type must be that same as the <see cref="Run"/> method</remarks>
     /// <remarks>Changing the <paramref name="input"/> or return type might break the Durable Function's deserialization</remarks>
     /// </summary>
     public static Task<int> StartActivityAsync(EnqueueMessagesForActorInput input, TaskOrchestrationContext context, TaskOptions? options)
     {
         return context.CallActivityAsync<int>(
-            nameof(EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity),
+            nameof(EnqueueWholesaleResultsForTotalAmountsActivity),
             input,
             options: options);
     }
 
-    [Function(nameof(EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity))]
+    [Function(nameof(EnqueueWholesaleResultsForTotalAmountsActivity))]
     public Task<int> Run([ActivityTrigger] EnqueueMessagesForActorInput input)
     {
-        var query = new WholesaleMonthlyAmountPerChargeQuery(
+        var query = new WholesaleTotalAmountQuery(
             _logger,
             _wholesaleResultEnumerator.EdiDatabricksOptions,
             input.GridAreaOwners,
@@ -63,10 +62,10 @@ public class EnqueueWholesaleResultsForMonthlyAmountPerChargesActivity(
 
     protected override Task EnqueueAndCommitWholesaleResult<TOutgoingMessage>(IOutgoingMessagesClient outgoingMessagesClient, TOutgoingMessage outgoingMessageDto)
     {
-        return outgoingMessageDto is WholesaleMonthlyAmountPerChargeMessageDto monthlyAmountPerChargeMessageDto
-            ? outgoingMessagesClient.EnqueueAndCommitAsync(monthlyAmountPerChargeMessageDto, CancellationToken.None)
+        return outgoingMessageDto is WholesaleTotalAmountMessageDto totalAmountMessageDto
+            ? outgoingMessagesClient.EnqueueAndCommitAsync(totalAmountMessageDto, CancellationToken.None)
             : throw new ArgumentException(
-                $"The outgoing message dto is not of the expected type {typeof(WholesaleMonthlyAmountPerChargeMessageDto).FullName}",
+                $"The outgoing message dto is not of the expected type {typeof(WholesaleTotalAmountMessageDto).FullName}",
                 nameof(outgoingMessageDto));
     }
 }
