+++ source/B2BApi/Functions/EnqueueMessages/BRS_021_023/Activities/GetActorsForWholesaleResultsForAmountPerChargesActivity.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Model;
+using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
 using Energinet.DataHub.EDI.MasterData.Interfaces;
 using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Databricks.WholesaleResults.Queries;
 using Microsoft.Azure.Functions.Worker;
@@ -21,36 +21,36 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using EventId = Energinet.DataHub.EDI.BuildingBlocks.Domain.Models.EventId;
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Activities;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Activities;
 
-public class GetActorsForWholesaleResultsForMonthlyAmountPerChargesActivity(
-    ILogger<GetActorsForWholesaleResultsForMonthlyAmountPerChargesActivity> logger,
+public class GetActorsForWholesaleResultsForAmountPerChargesActivity(
+    ILogger<GetActorsForWholesaleResultsForAmountPerChargesActivity> logger,
     IServiceScopeFactory serviceScopeFactory,
     IMasterDataClient masterDataClient,
     WholesaleResultActorsEnumerator wholesaleResultActorsEnumerator)
 {
-    private readonly ILogger<GetActorsForWholesaleResultsForMonthlyAmountPerChargesActivity> _logger = logger;
+    private readonly ILogger<GetActorsForWholesaleResultsForAmountPerChargesActivity> _logger = logger;
     private readonly IServiceScopeFactory _serviceScopeFactory = serviceScopeFactory;
     private readonly IMasterDataClient _masterDataClient = masterDataClient;
     private readonly WholesaleResultActorsEnumerator _wholesaleResultActorsEnumerator = wholesaleResultActorsEnumerator;
 
     /// <summary>
-    /// Start an GetActorsForWholesaleResultsForMonthlyAmountPerCharges activity.
+    /// Start an GetActorsForWholesaleResultsForAmountPerCharges activity.
     /// <remarks>The <paramref name="input"/> type and return type must be that same as the <see cref="Run"/> method</remarks>
     /// <remarks>Changing the <paramref name="input"/> or return type might break the Durable Function's deserialization</remarks>
     /// </summary>
     public static Task<IReadOnlyCollection<string>> StartActivityAsync(EnqueueMessagesInput input, TaskOrchestrationContext context, TaskOptions? options)
     {
         return context.CallActivityAsync<IReadOnlyCollection<string>>(
-            nameof(GetActorsForWholesaleResultsForMonthlyAmountPerChargesActivity),
+            nameof(GetActorsForWholesaleResultsForAmountPerChargesActivity),
             input,
             options: options);
     }
 
-    [Function(nameof(GetActorsForWholesaleResultsForMonthlyAmountPerChargesActivity))]
+    [Function(nameof(GetActorsForWholesaleResultsForAmountPerChargesActivity))]
     public async Task<IReadOnlyCollection<string>> Run([ActivityTrigger] EnqueueMessagesInput input)
     {
-        var query = new WholesaleMonthlyAmountPerChargeQuery(
+        var query = new WholesaleAmountPerChargeQuery(
             _logger,
             _wholesaleResultActorsEnumerator.EdiDatabricksOptions,
             input.GridAreaOwners,
