+++ source/B2BApi/Functions/EnqueueMessages/BRS_021_023/EnqueueMessagesOrchestration.cs
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Activities;
-using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Model;
+using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Activities;
+using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
 using Energinet.DataHub.EDI.B2BApi.MetricTracker;
 using Energinet.DataHub.EDI.IntegrationEvents.Infrastructure.Model;
 using Microsoft.ApplicationInsights;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.DurableTask;
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023;
 
 internal class EnqueueMessagesOrchestration(TelemetryClient telemetryClient)
 {
