+++ source/B2BApi/Functions/EnqueueMessages/BRS_021_023/Model/SendMessagesEnqueuedInput.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.Model;
+namespace Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_021_023.Model;
 
 /// <summary>
 /// An immutable input to start the send messages enqueued activity.
